requirements.txt
kmodes==0.10.2
matplotlib==3.2.1
statsmodels==0.11.1
pandas==1.1.3
pmdarima==1.7.1
fbprophet==0.7.1
tqdm==4.45.0
numpy==1.18.2
python_dateutil==2.8.1
PyYAML==5.2
scikit_learn==0.23.2
